Extract handleSearch function from SearchProductWithBarcode (#5)
Summary:
We extracted the `handleSearch` function into the utils folder to make it more reusable. In order for it to function properly, we moved the state definitions up a level into `App.js` and passed it as props into the `SearchProductWithBarcode` component.

To-do:
Update the `handleSearch` function so it can use other functions besides `getProductByBarcode` to make it truly reusable.