Create H1 component (#6)
Summary:
I created a H1 component that can be reused in the app. It takes in a `title` prop, which allows the title to be displayed and also has a test to check if the component renders. Although there is currently no design, I believe preparing for when there is will make for an easier switch.

To-do:
There is no design as of yet, but when there is we can style this component. This would probably be through a `H1.css` stylesheet or a React JSS styles object. We should probably also allow the component to have custom styles by adding a `customStyles` prop.