Create P component (#7)
Summary:
I created a `P` component that can be reused in the app. It takes in `text` and `children` props, which allows the text to be displayed and have things nested in it just in case. There are also has test to check if the component renders and can accept a child.

To-do:
Similar to the `H1` component, this one needs to be styled as well.