Create SearchBar component (#8)
Summary:
I created a `SearchBar` component, which is essentially a form with a hidden label, an input field and a submit button. It also accepts two functions, `handleSearch` and `handleClick` I added a bit of styling and made it responsive just for fun, but still waiting on the design for the official styling. There are tests to check all of the elements are present.

To-do:
Check if both `handleClick` and `handleSearch` are needed.