Create ProductCard (#9)
Summary:
I created a component that consists of `img`, `h1` and `p` elements, and together they make up the `ProductCard`. There are also tests that check that each element displays. However, for the `<img />` in the test, there is no `src`, so I've just checked the element exists in the document.

To do:
Again, there is some styling left to do.