Create SearchByName component (#13)
Summary:
We've created a `SearchByName` component that allows users to search for products by their name and then displays the result in a grid of `ProductCard`s. The component brings together `ProductCard`, `ProductGrid` and `SearchBar`.

The `useEffect()` section is adapted from a solution by someone (Dan) in a Github thread (check PR for link).

We've also got some conditional rendering going on line 35 (`userInput != '' &&`... ) because the API seems to be returning some result even when the search bar is empty. So, in order not to display them, I used conditional rendering.

To do:
Should've TDD'd this, so definitely need some tests. 👀
Also, would love to combine this with `SearchProductByBarcode`'s functionality, so we have a search bar that can do both 😉